requirements.txt
asgiref==3.5.2
black==22.6.0
boto3==1.24.63
botocore==1.27.63
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
dj-database-url==1.0.0
Django==4.0.6
django-cors-headers==3.13.0
django-heroku==0.3.1
django-rest-knox==4.2.0
django-storages==1.13.1
djangorestframework==3.13.1
drf-yasg==1.21.3
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
mysqlclient==2.1.1
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
stripe==3.5.0
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.11
whitenoise==6.2.0
